requirements.txt
torch>=2.3.1
torchvision>=0.18.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
streamlit-drawable-canvas>=0.9.3
opencv-python>=4.10.0.84
stability-sdk>=0.8.6